rename folders to be lower case
