+++ source/nodejs/adaptivecards-fabric/src/components/Fabric.tsx
