+++ source/nodejs/adaptivecards-fabric/src/components/inputs/InputToggleFabric.tsx
