+++ source/nodejs/adaptivecards-fabric/src/index.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-export * from "./Components";
-export * from "./Actions";
-export * from "./Utils/shared";
+export * from "./components";
+export * from "./actions";
+export * from "./utils/shared";
