+++ source/nodejs/adaptivecards-fabric/src/utils/shared.tsx
